Continue when hotend temp above target when probing
This commit prevents the probing procedure to wait for the hotend
temperature to settle if its above the target temperature.
This is desired, if for example, the first layer failed and the
print must be restarted.

We would be forced to wait until the temperature has fallen,
depending on the chamber environment and hotend,
this could take several minutes.

Instead we ignore if its above and continue with probing, this saves
time, especially when prototyping. It shouldn't be a problem anyways.